requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
cffi==1.15.0
Django==4.0.3
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.0.1
psycopg2==2.9.3
pycparser==2.21
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
